@@ -3,12 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.1.2",
+    "@fortawesome/free-brands-svg-icons": "^6.1.2",
+    "@fortawesome/free-regular-svg-icons": "^6.1.2",
+    "@fortawesome/free-solid-svg-icons": "^6.1.2",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "axios": "^0.27.2",
     "http-proxy-middleware": "^2.0.6",
+    "i": "^0.3.7",
     "medium-editor": "^5.23.3",
+    "npm": "^8.16.0",
     "quill-image-resize": "^3.0.9",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-icons": "^4.4.0",
     "react-identicons": "^1.1.7",
     "react-medium-editor": "^1.8.1",
     "react-modal": "^3.15.1",
