@@ -1,6 +1,6 @@
-import Head from "./page/Head";
 import React from "react";
-import TodoApp from "./page/main/TodoApp";
+import Head from "./page/Head";
+// import TodoApp from "./page/main/TodoApp";
 
 function App() {
   return (
