@@ -1,12 +1,16 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
-import Head from "./Head";
+import Header from "./main/Header";
+import Sidebar from "./main/Sidebar";
 
 function Main() {
   return (
     <div>
-      <Head />
-      <Outlet />
+      <Header />
+      <div style={{ display: "grid", gridTemplateColumns: "0.5fr 2fr" }}>
+        <Sidebar />
+        <Outlet />
+      </div>
     </div>
   );
 }
