@@ -1,58 +1,17 @@
 import React, { useState, useEffect } from "react";
 import Modal from "./Modal";
-import DelModal from "./DelModal";
-import axios from "axios";
-import SucessModal from "./SucessModal";
+import { getList, openInfoModal } from "../../api/documentApi";
+import Page from "./Page";
 
-const getList = (setList, page) => {
-  axios.get("http://localhost:8080/api/documents/user/2").then((res) => {
-    setList(res.data);
-    console.log(res.data);
-  });
-};
-
-const openModal = (setModalOpen, documentNo, setDocument) => {
-  setModalOpen(true);
-  axios
-    .get("http://localhost:8080/api/document/" + documentNo)
-    .then((res) => setDocument(res.data));
-};
-const closeModal = (setModalOpen) => {
-  setModalOpen(false);
-};
-
-const deleteFile = (setModalOpen, setModalOpen2, setModalOpen3, documentNo) => {
-  axios
-    .delete("http://localhost:8080/api/document/" + documentNo)
-    .then(setModalOpen(false));
-  setModalOpen2(false);
-  setModalOpen3(true);
-};
-
-const openModal2 = (setModalOpen2) => {
-  setModalOpen2(true);
-};
-const closeModal2 = (setModalOpen2) => {
-  setModalOpen2(false);
-};
-const closeModal3 = (setModalOpen3, setModalOpen) => {
-  setModalOpen3(false);
-  setModalOpen(false);
-};
 function MyBox() {
-  // useState를 사용하여 open상태를 변경한다. (open일때 true로 만들어 열리는 방식)
-  const [modalOpen, setModalOpen] = useState(false);
-  const [modalOpen2, setModalOpen2] = useState(false);
-  const [modalOpen3, setModalOpen3] = useState(false);
-
+  const [infoModalOpen, setInfoModalOpen] = useState(false);
   const [document, setDocument] = useState("");
-
   const [page, setPage] = useState(1);
   const [list, setList] = useState("");
 
   useEffect(() => {
     getList(setList, page);
-  }, [page, modalOpen3]);
+  }, [page, infoModalOpen]);
   return (
     <React.Fragment>
       <div>
@@ -80,8 +39,8 @@ function MyBox() {
 
                     <td
                       onClick={() =>
-                        openModal(
-                          setModalOpen,
+                        openInfoModal(
+                          setInfoModalOpen,
                           docs.documentNo.documentNo,
                           setDocument
                         )
@@ -101,77 +60,13 @@ function MyBox() {
           </tbody>
         </table>
       </div>
-      <div>
-        <button
-          onClick={() =>
-            setPage(list.start - list.size >= 1 ? list.start - list.size : page)
-          }
-        >
-          이전
-        </button>
-
-        {list &&
-          list.pageList.map((listPage) => {
-            return (
-              <span
-                className={page === listPage ? "pageNum" : "pageOut"}
-                onClick={() => {
-                  setPage(listPage);
-                }}
-                key={listPage}
-              >
-                {listPage}
-              </span>
-            );
-          })}
-        <button
-          onClick={() =>
-            setPage(
-              list.start + list.size > list.totalPage
-                ? page
-                : list.start + list.size
-            )
-          }
-        >
-          다음
-        </button>
-      </div>
+      <Page list={list} page={page} setPage={setPage} />
 
       <Modal
-        open={modalOpen}
-        close={() => closeModal(setModalOpen)}
-        header={document.originalName}
-        down={document.filePath}
-        open2={() => openModal2(setModalOpen2)}
-      >
-        <main>
-          <div>{document.content}</div>
-        </main>
-      </Modal>
-      <DelModal
-        open={modalOpen2}
-        close={() => closeModal2(setModalOpen2)}
-        del={() =>
-          deleteFile(
-            setModalOpen,
-            setModalOpen2,
-            setModalOpen3,
-            document.documentNo
-          )
-        }
-      >
-        <main>
-          <div>삭제하시겠습니까?</div>
-        </main>
-      </DelModal>
-      <SucessModal
-        open={modalOpen3}
-        close={() => closeModal3(setModalOpen3, setModalOpen)}
-      >
-        <main>
-          <div>삭제 완료</div>
-        </main>
-      </SucessModal>
+        open={infoModalOpen}
+        document={document}
+        infoModalOpen={setInfoModalOpen}
+      />
     </React.Fragment>
   );
 }
