@@ -3,20 +3,23 @@ import {
   deleteUserWorkspace,
   getUserWorkspace,
 } from "../../api/workspaceUserApi";
+import { getUser } from "../../component/getUser/getUser";
 
 function WorkspaceList() {
   const [workspaces, setWorkspace] = useState([]);
   useEffect(() => {
     getUserWorkspace(1, setWorkspace);
   }, []);
-  console.log(workspaces);
+  console.log(getUser());
+
+  // console.log(JSON.parse(window.localStorage.getItem("event")));
   return (
     <div>
       {workspaces.map((workspace) => (
         <div className="workspaceList" key={workspace.workspaceNo.workspaceNo}>
           <span>{workspace.workspaceNo.title}</span>
           <span>{workspace.workspaceNo.master.name}</span>
-          <a href={`/main/document/${workspace.workspaceNo.workspaceNo}`}>
+          <a href={`/main/document?room=${workspace.workspaceNo.workspaceNo}`}>
             <button>입장</button>
           </a>
           <button
